Major Code Refactor
This update is in the middle of a major code refactor. Dictionaries were turned into classes, the only dictionary left is the locations dictionary, and everything else is being turned into functions and placed into corresponding classes. This is probably the biggest update, but hopefully this will make the development of new mechanics and game play much easier in the future. 

On the plus side, the code after the refactor is looking really good, and I'm learning a lot about python along the way.